refactor(filter): seperate components
