+++ src/components/DesktopFilter/Filter.storybook.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { CSEFilterParams, CSEFilterOptions } from '../../models/filter';
-import { Filter } from '.';
+import { DesktopFilter } from '.';
 import { filterOptionsGS, filterOptionsGSC, filterOptionsGSCT } from './__mocks__';
 import { storiesOf } from '@storybook/react';
 import centered from '@storybook/addon-centered';
@@ -13,48 +13,48 @@ const callback = (newFilter: CSEFilterParams) => {
 storiesOf('Filter', module)
   .addDecorator(centered)
   .add('default', () => (
-    <Filter params={{ grades: [] }} options={filterOptionsGS} onUpdate={callback} />
+    <DesktopFilter params={{ grades: [] }} options={filterOptionsGS} onUpdate={callback} />
   ))
   .add('selected grade', () => (
-    <Filter params={{ grades: ['ms'] }} options={filterOptionsGS} onUpdate={callback} />
+    <DesktopFilter params={{ grades: ['ms'] }} options={filterOptionsGS} onUpdate={callback} />
   ))
   .add('selected subject', () => (
-    <Filter
+    <DesktopFilter
       params={{ grades: [], subject: 'math' }}
       options={filterOptionsGSC}
       onUpdate={callback}
     />
   ))
   .add('selected subject, claim', () => (
-    <Filter
+    <DesktopFilter
       params={{ grades: [], subject: 'math', claim: 'm1' }}
       options={filterOptionsGSCT}
       onUpdate={callback}
     />
   ))
   .add('selected subject, claim, target', () => (
-    <Filter
+    <DesktopFilter
       params={{ grades: [], subject: 'math', claim: 'm1', target: 'pf' }}
       options={filterOptionsGSCT}
       onUpdate={callback}
     />
   ))
   .add('selected grades, subject, claim, target', () => (
-    <Filter
+    <DesktopFilter
       params={{ grades: ['ms', 'hs'], subject: 'math', claim: 'm1', target: 'pf' }}
       options={filterOptionsGSCT}
       onUpdate={callback}
     />
   ))
   .add('invalid', () => (
-    <Filter
+    <DesktopFilter
       params={{ grades: ['ms'], claim: 'm1', target: 'pf' }}
       options={filterOptionsGS}
       onUpdate={callback}
     />
   ))
   .add('invalid 2', () => (
-    <Filter
+    <DesktopFilter
       params={{ grades: ['ms'], subject: 'math', target: 'pf' }}
       options={filterOptionsGSC}
       onUpdate={callback}
