+++ src/components/DesktopFilter/index.spec.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import { Filter } from '.';
+import { DesktopFilter } from '.';
 import { filterOptionsGS, filterOptionsGSC, filterOptionsGSCT } from './__mocks__';
 
 describe('Filter', () => {
   it('calls callback on reset button clicked', () => {
     const callback = jest.fn();
     const component = mount(
-      <Filter params={{ grades: [] }} options={filterOptionsGS} onUpdate={callback} />
+      <DesktopFilter params={{ grades: [] }} options={filterOptionsGS} onUpdate={callback} />
     );
 
     component
