+++ src/components/FilterHelper/index.ts
@@ -4,7 +4,7 @@ import {
   FilterOptionModel,
   SearchBaseModel
 } from '@osu-cass/sb-components';
-import { CSEAdvancedFilterModels } from './';
+import { CSEAdvancedFilterModels } from '../FilterProps';
 import { CSEFilterOptions, CSEFilterParams } from '../../models/filter';
 
 export function createFilters(
