+++ src/components/SearchPage/index.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import { SearchBar } from '../SearchBar';
 import { FilterItemList } from '../FilterItemList';
-import { Filter } from '../Filter/DesktopFilter';
 import { CSEFilterOptions, CSEFilterParams, CSESearchQuery } from '../../models/filter';
 import { Styles } from '../../constants';
 import { FilterContianer } from '../FilterContainer';
-import { Message, ErrorMessage } from '../Filter/Messages';
+import { Message, ErrorMessage } from '../FilterMessage';
 import { FilterType, ErrorBoundary } from '@osu-cass/sb-components';
 import { ISearchClient } from '../../clients/search';
 import { IClaim } from '../../models/claim';
